Fix JSON export function in M365DSCReport.psm1
Implemented a temporary fix for the New-M365DSCConfigurationToJSON function. There is a logic error with the current function and it will always export to null. I implemented a solution utilizing substrings and just subtracted the amount of characters in the last two lines of the export. 

This somewhat hardcoded solution will work for the time being, but I plan on writing a more future-proof solution soon.