Update MSFT_IntuneAppProtectionPolicyAndroid.psm1
working on the update command - modifed function to get properties of different commands and set the parameters accordingly.  Moved paams code outside for the new if statement.  Had to remove Apps and Assignments to get it to function with Updates - we also need to specify the ID. This needs more work but it will amend an existing policy now