Update MSFT_IntuneAppProtectionPolicyAndroid.psm1
Trying to get test to work correctly when items are omitted.  I could just make them mandatory, of course, but I'd prefer it to make a comparison using blank values.  otherwise omitting a result wouldn't set it back to nothing on an existing config (maybe that's the desired outcome, of course...)