Unconditionally remove ReplyURLs (and friends) from $currentParameters
Whenever an application is both present in the tenant and in the DSC side then "Update-MgApplication @currentParameters" is always called but if $ReplyURLs is empty then it errors out because it will try to use it as parameter for Update-MgApplication which doesn't exist. To fix this just remove $ReplyURLs and friends from $currentParameters.