Abort extraction if all components are invalid or if resource files were not found
When calling Export-M365DSCConfiguration, only with a list of invalid components, then Start-M365DSCConfigurationExtract is called internally without validating that list. This means it tries to export as much as it can and depending on the accesses required may start giving access denied error messages. Additionally be careful to not proceed with extraction if in the meantime user deletes all resource files, this should not happen but it doesn't hurt to have the check.