Dedup some code on O365SiteGroup
Avoid redefining $SiteGroupSettings always to the same value, just define it once, and call it as is on Set-PnPGroup, except if it's to be used on remove the SiteGroup, in this case just remove the Owner key/value. Additionally to keep the order of setting the group owner first, and then the group permissions add a new object ($GroupPermissionsParameters) to keep track of which parameters are required and then finally call Set-PnPGroup/Set-PnPGroupPermissions in the right order.