Fix typo in variable and adapt surrounding code
Fix typo in variable, currently this is only a nuisance since even if Identity and Owner of SiteGroup is the same it tries to update it anyway to the same values, so make it clear that if they are the same updating the SiteGroup is not required. While here adapt names of variables surrounding this code.