CustomBrowserDisplayName, CustomBrowserPackageId beguin and set AppGroupType to export as string
I've begun work on the custombrowser id settings but not completed yet - need to process as a complex parameter as they can set non viable values in the gui.  I've begun a function Function set-ManagedBrowserValues but it doesn't do anything yet. Also fixed an issue where AppGroupType isn't exporting as a string